# [6.0]optimize header value code