```shell
// without this patch benchmarks
wrk -t4 -c500 -d30s -s script.lua http://127.0.0.1/
Running 30s test @ http://127.0.0.1/
  4 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.51ms    1.92ms  23.82ms   72.97%
    Req/Sec    25.95k     1.68k   40.07k    74.08%
  3101765 requests in 30.08s, 485.12MB read
Requests/sec: 103102.72
Transfer/sec:     16.13MB
```

```shell
// with this patch benchmarks
wrk -t4 -c500 -d30s -s script.lua http://127.0.0.1/
Running 30s test @ http://127.0.0.1/
  4 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.41ms    1.92ms  27.18ms   72.86%
    Req/Sec    26.43k     1.66k   31.30k    72.49%
  3162096 requests in 30.08s, 494.56MB read
Requests/sec: 105121.75
Transfer/sec:     16.44MB
```